+++ NEWS.md
@@ -138,6 +138,8 @@ Standard library changes
   (callable via `cholesky[!](A, RowMaximum())`) ([#54619]).
 * The number of default BLAS threads now respects process affinity, instead of
   using total number of logical threads available on the system ([#55574]).
+* A new function `zeroslike` is added that is used to generate the zero elements for matrix-valued banded matrices.
+  Custom array types may specialize this function to return an appropriate result. ([#55252])
 
 #### Logging